only able to delete if user or admin

+++ src/buttons/delete.ts
@@ -1,5 +1,19 @@
-import { ButtonInteraction, Message } from "discord.js";
+import { ButtonInteraction, Message, Permissions } from "discord.js";
 
 export default async function (interaction: ButtonInteraction) {
+  const userThatCanDelete = await interaction.client.users.fetch(
+    interaction.customId.split("-")[1]
+  );
+
+  if (
+    interaction.user.id != userThatCanDelete.id &&
+    !interaction.memberPermissions.has(Permissions.FLAGS.ADMINISTRATOR)
+  )
+    return await interaction.reply({
+      content:
+        "Only the user who converted that tiktok or someone with administrator privileges is able to delete that message.",
+      ephemeral: true,
+    });
+
   await (interaction.message as Message).delete();
 }
